Fix temperature measuring
Temperature raw data is a 20 bit vector stored in a signed 32 bit
integer. When putting the MSB, LSB and XLSB together, the variable
storing the raw data should be masked with 0xFFFFF in order to clear
leading bits (which in this case were all '1'), as a result data read
were negative, and calibration was giving wrong values.
Same goes with pressure.